Set user_oid in session and PageManager
We can now show User profile link on all pages and easier handle the
current user in other views.

For example easily finding out if a password key is connected to a user
by looking up the user by its oid which is already available to us.

Before only the username was available in request.session.